==> code/01_library.R <==

################          LIBRARIES USED         ################

library(sp)
library(raster)
library(sf)
library(maps)
library(mapdata)
library(maptools)
library(rgdal)
library(gdalUtils) 
library(gdalUtilities)
library(gpclib) 
library(spatstat)
library(RGtk2)
library(MODIStsp)
library(MODIS)
library(ggplot2)
library(ggmap)
library(lubridate)
library(rgeos)
library(reshape2)
library(dplyr)
library(lmtest)
library(tm)
library(csv)
library(csvread)
library(R.utils)
library(lattice)
library(caret)
library(RStoolbox)
library(RColorBrewer)
library(openair)
library(gridExtra)
library(printr)
library(corrplot)
library(randomForest)
library(xlsx)
library(tdr)
library(tidyverse)
library(ggExtra)
library(ggpp)
library(ggmap)
library(ggpubr)
library(readr)
library(tidyverse)
library(httr)
library(jsonlite)
library(ggmap)
library(gepaf)
library(dplyr)
library(ncdf4)
library(dplyr)
library(sp)
library(raster)
library(dplyr)
library(purrr)
library(stringr)
library(lubridate)
library(foreach)
library(googlePolylines)
